refactor alert filters
