+++ src/py42/sdk/queries/alerts/filters/alert_filter.py
@@ -1,119 +1,41 @@
-from py42.choices import Choices
-from py42.sdk.queries.query_filter import create_filter_group
-from py42.sdk.queries.query_filter import create_query_filter
-from py42.sdk.queries.query_filter import QueryFilterStringField
-from py42.sdk.queries.query_filter import QueryFilterTimestampField
-from py42.util import MICROSECOND_FORMAT
-from py42.util import parse_timestamp_to_microseconds_precision
+from py42.choices import Choices as _Choices
+from py42.sdk.queries.alerts.util import (
+    AlertQueryFilterStringField as _AlertQueryFilterStringField,
+)
+from py42.sdk.queries.alerts.util import (
+    AlertQueryFilterTimestampField as _AlertQueryFilterTimestampField,
+)
+from py42.sdk.queries.query_filter import (
+    QueryFilterStringField as _QueryFilterStringField,
+)
 
 
-def create_contains_filter_group(term, value):
-    """Creates a :class:`~py42.sdk.queries.query_filter.FilterGroup` for filtering results
-    where the value with key ``term`` contains the given value. Useful for creating ``CONTAINS``
-    filters that are not yet supported in py42 or programmatically crafting filter groups.
-
-    Args:
-        term: (str): The term of the filter, such as ``actor``.
-        value (str): The value used to match on.
-
-    Returns:
-        :class:`~py42.sdk.queries.query_filter.FilterGroup`
-    """
-
-    filter_list = [create_query_filter(term, "CONTAINS", value)]
-    return create_filter_group(filter_list, "AND")
-
-
-def create_not_contains_filter_group(term, value):
-    """Creates a :class:`~py42.sdk.queries.query_filter.FilterGroup` for filtering results
-    where the value with key ``term`` does not contain the given value. Useful for creating
-    ``DOES_NOT_CONTAIN`` filters that are not yet supported in py42 or programmatically
-    crafting filter groups.
-
-    Args:
-        term: (str): The term of the filter, such as ``actor``.
-        value (str): The value used to exclude on.
-
-    Returns:
-        :class:`~py42.sdk.queries.query_filter.FilterGroup`
-    """
-
-    filter_list = [create_query_filter(term, "DOES_NOT_CONTAIN", value)]
-    return create_filter_group(filter_list, "AND")
-
-
-class AlertQueryFilterStringField(QueryFilterStringField):
-    @classmethod
-    def contains(cls, value):
-        """Creates a :class:`~py42.sdk.queries.query_filter.FilterGroup` for filtering
-        results where the value with key ``self._term`` contains the given value. Useful
-        for creating ``CONTAINS`` filters that are not yet supported in py42 or programmatically
-        crafting filter groups.
-
-        Args:
-            value (str): The value used to match on.
-
-        Returns:
-            :class:`~py42.sdk.queries.query_filter.FilterGroup`
-        """
-
-        return create_contains_filter_group(cls._term, value)
-
-    @classmethod
-    def not_contains(cls, value):
-        """Creates a :class:`~py42.sdk.queries.query_filter.FilterGroup` for filtering
-        results where the value with key ``self._term`` does not contain the given value.
-        Useful for creating ``DOES_NOT_CONTAIN`` filters that are not yet supported in py42
-        or programmatically crafting filter groups.
-
-        Args:
-            value (str): The value used to exclude on.
-
-        Returns:
-            :class:`~py42.sdk.queries.query_filter.FilterGroup`
-        """
-
-        return create_not_contains_filter_group(cls._term, value)
-
-
-class AlertQueryFilterTimestampField(QueryFilterTimestampField):
-    """Helper class for creating alert filters where the search value is a timestamp."""
-
-    @staticmethod
-    def _parse_timestamp(value):
-        return parse_timestamp_to_microseconds_precision(value)
-
-    @staticmethod
-    def _convert_datetime_to_timestamp(value):
-        return value.strftime(MICROSECOND_FORMAT)
-
-
-class DateObserved(AlertQueryFilterTimestampField):
+class DateObserved(_AlertQueryFilterTimestampField):
     """Class that filters alerts based on the timestamp the alert was triggered."""
 
     _term = "createdAt"
 
 
-class Actor(AlertQueryFilterStringField):
+class Actor(_AlertQueryFilterStringField):
     """Class that filters alerts based on the username that originated the event(s) that
     triggered the alert."""
 
     _term = "actor"
 
 
-class RuleName(AlertQueryFilterStringField):
+class RuleName(_AlertQueryFilterStringField):
     """Class that filters alerts based on rule name."""
 
     _term = "name"
 
 
-class RuleId(QueryFilterStringField):
+class RuleId(_QueryFilterStringField):
     """Class that filters alerts based on rule identifier."""
 
     _term = "ruleId"
 
 
-class RuleSource(QueryFilterStringField, Choices):
+class RuleSource(_QueryFilterStringField, _Choices):
     """Class that filters alerts based on rule source.
 
     Available options are:
@@ -129,7 +51,7 @@ class RuleSource(QueryFilterStringField, Choices):
     HIGH_RISK_EMPLOYEE = "High Risk Employee"
 
 
-class RuleType(QueryFilterStringField, Choices):
+class RuleType(_QueryFilterStringField, _Choices):
     """Class that filters alerts based on rule type.
 
     Available options are:
@@ -145,13 +67,13 @@ class RuleType(QueryFilterStringField, Choices):
     FILE_TYPE_MISMATCH = "FedFileTypeMismatch"
 
 
-class Description(AlertQueryFilterStringField):
+class Description(_AlertQueryFilterStringField):
     """Class that filters alerts based on rule description text."""
 
     _term = "description"
 
 
-class Severity(QueryFilterStringField, Choices):
+class Severity(_QueryFilterStringField, _Choices):
     """Class that filters alerts based on severity.
 
     Available options are:
@@ -170,7 +92,7 @@ class Severity(QueryFilterStringField, Choices):
     LOW = "LOW"
 
 
-class AlertState(QueryFilterStringField, Choices):
+class AlertState(_QueryFilterStringField, _Choices):
     """Class that filters alerts based on alert state.
 
     Available options are:
